--- controllers/contacts.js
@@ -1,60 +1,0 @@
-const { Contacts } = require("../models/contacts");
-
-const { HttpError, ctrlWrapper } = require("../helpers");
-
-const listContacts = async (req, res) => {
-  const result = await Contacts.find({}, "-createdAt -updatedAt");
-  res.json(result);
-};
-
-const getContactById = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contacts.findById(id);
-  if (!result) {
-    throw HttpError(404, "not found");
-  }
-  res.json(result);
-};
-
-const addContact = async (req, res) => {
-  const result = await Contacts.create(req.body);
-  res.status(201).json(result);
-};
-
-const updateContact = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contacts.findByIdAndUpdate(id, req.body, { new: true });
-  if (!result) {
-    throw HttpError(404, "not found");
-  }
-  res.json(result);
-};
-
-const updateFavorite = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contacts.findByIdAndUpdate(id, req.body, { new: true });
-  if (!result) {
-    throw HttpError(404, "not found");
-  }
-  res.json(result);
-};
-
-const removeContact = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contacts.findByIdAndRemove(id);
-  if (!result) {
-    throw HttpError(404, "not found");
-  }
-  res.json({
-    message: "contact deleted",
-  });
-};
-
-module.exports = {
-  listContacts: ctrlWrapper(listContacts),
-  getContactById: ctrlWrapper(getContactById),
-  addContact: ctrlWrapper(addContact),
-  updateContact: ctrlWrapper(updateContact),
-  updateFavorite: ctrlWrapper(updateFavorite),
-  removeContact: ctrlWrapper(removeContact),
-};
