+++ controllers/contacts/addContact.js
@@ -1,0 +1,8 @@
+const { Contacts } = require("../../models/contacts");
+
+const addContact = async (req, res) => {
+  const result = await Contacts.create(req.body);
+  res.status(201).json(result);
+};
+
+module.exports = addContact;
