+++ controllers/contacts/getAll.js
@@ -1,0 +1,8 @@
+const { Contacts } = require("../../models/contacts");
+
+const listContacts = async (req, res) => {
+  const result = await Contacts.find({}, "-createdAt -updatedAt");
+  res.json(result);
+};
+
+module.exports = listContacts;
