+++ controllers/contacts/index.js
@@ -1,0 +1,16 @@
+const { ctrlWrapper } = require("../../helpers");
+const listContacts = require("./getAll");
+const getContactById = require("./getById");
+const addContact = require("./addContact");
+const updateContact = require("./updateContact");
+const updateFavorite = require("./updateFavorite");
+const removeContact = require("./removeContact");
+
+module.exports = {
+  listContacts: ctrlWrapper(listContacts),
+  getContactById: ctrlWrapper(getContactById),
+  addContact: ctrlWrapper(addContact),
+  updateContact: ctrlWrapper(updateContact),
+  updateFavorite: ctrlWrapper(updateFavorite),
+  removeContact: ctrlWrapper(removeContact),
+};
