+++ controllers/index.js
@@ -1,0 +1,3 @@
+const controllers = require("./contacts");
+
+module.exports = controllers;
