+++ models/index.js
@@ -1,0 +1,3 @@
+const Contact = require("./contacts");
+
+module.exports = { Contact };
