+++ schemas/addInfo.js
@@ -7,6 +7,4 @@ const addSchema = Joi.object({
   favorite: Joi.boolean(),
 });
 
-module.exports = {
-  addSchema,
-};
+module.exports = addSchema;
