+++ schemas/index.js
@@ -1,0 +1,6 @@
+const addSchema = require("./addInfo");
+const updateFavorite = require("./updateFavorite");
+
+const schema = { addSchema, updateFavorite };
+
+module.exports = schema;
